Switch to a proper builder for MockRetrofit.
This allows passing an explicit executor for test cases.
