+++ retrofit-mock/src/main/java/retrofit2/mock/BehaviorDelegate.java
@@ -20,6 +20,7 @@ import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
 import java.lang.reflect.Type;
+import java.util.concurrent.ExecutorService;
 import retrofit2.Call;
 import retrofit2.CallAdapter;
 import retrofit2.Retrofit;
@@ -34,11 +35,14 @@ import retrofit2.Retrofit;
 public final class BehaviorDelegate<T> {
   private final Retrofit retrofit;
   private final NetworkBehavior behavior;
+  private final ExecutorService executor;
   private final Class<T> service;
 
-  BehaviorDelegate(Retrofit retrofit, NetworkBehavior behavior, Class<T> service) {
+  BehaviorDelegate(Retrofit retrofit, NetworkBehavior behavior, ExecutorService executor,
+      Class<T> service) {
     this.retrofit = retrofit;
     this.behavior = behavior;
+    this.executor = executor;
     this.service = service;
   }
 
@@ -48,8 +52,7 @@ public final class BehaviorDelegate<T> {
 
   @SuppressWarnings("unchecked") // Single-interface proxy creation guarded by parameter safety.
   public T returning(Call<?> call) {
-    final Call<?> behaviorCall =
-        new BehaviorCall<>(behavior, retrofit.client().getDispatcher().getExecutorService(), call);
+    final Call<?> behaviorCall = new BehaviorCall<>(behavior, executor, call);
     return (T) Proxy.newProxyInstance(service.getClassLoader(), new Class[] { service },
         new InvocationHandler() {
           @Override
