+++ retrofit-mock/src/main/java/retrofit2/mock/MockRetrofit.java
@@ -15,23 +15,20 @@
  */
 package retrofit2.mock;
 
+import java.util.concurrent.Executor;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import retrofit2.Retrofit;
 
 public final class MockRetrofit {
-  public static MockRetrofit create(Retrofit retrofit) {
-    return create(retrofit, NetworkBehavior.create());
-  }
-
-  public static MockRetrofit create(Retrofit retrofit, NetworkBehavior behavior) {
-    return new MockRetrofit(retrofit, behavior);
-  }
-
   private final Retrofit retrofit;
   private final NetworkBehavior behavior;
+  private final ExecutorService executor;
 
-  public MockRetrofit(Retrofit retrofit, NetworkBehavior behavior) {
+  MockRetrofit(Retrofit retrofit, NetworkBehavior behavior, ExecutorService executor) {
     this.retrofit = retrofit;
     this.behavior = behavior;
+    this.executor = executor;
   }
 
   public Retrofit retrofit() {
@@ -42,8 +39,41 @@ public final class MockRetrofit {
     return behavior;
   }
 
+  public Executor backgroundExecutor() {
+    return executor;
+  }
+
   @SuppressWarnings("unchecked") // Single-interface proxy creation guarded by parameter safety.
   public <T> BehaviorDelegate<T> create(Class<T> service) {
-    return new BehaviorDelegate<>(retrofit, behavior, service);
+    return new BehaviorDelegate<>(retrofit, behavior, executor, service);
+  }
+
+  public static final class Builder {
+    private final Retrofit retrofit;
+    private NetworkBehavior behavior;
+    private ExecutorService executor;
+
+    public Builder(Retrofit retrofit) {
+      if (retrofit == null) throw new NullPointerException("retrofit == null");
+      this.retrofit = retrofit;
+    }
+
+    public Builder networkBehavior(NetworkBehavior behavior) {
+      if (behavior == null) throw new NullPointerException("behavior == null");
+      this.behavior = behavior;
+      return this;
+    }
+
+    public Builder backgroundExecutor(ExecutorService executor) {
+      if (executor == null) throw new NullPointerException("executor == null");
+      this.executor = executor;
+      return this;
+    }
+
+    public MockRetrofit build() {
+      if (behavior == null) behavior = NetworkBehavior.create();
+      if (executor == null) executor = Executors.newCachedThreadPool();
+      return new MockRetrofit(retrofit, behavior, executor);
+    }
   }
 }
